The PaginationInfo  ,BrandMapProfile , GetAllBrandQueryRequestHandler have been refactored.

+++ src/Common/RentACarNow.Common.Models/PaginationInfo.cs
@@ -4,8 +4,8 @@
     {
         public int PageNumber { get; set; }
         public int PageItemSize { get; set; }
-        public int TotalItemCount { get; set; }
-        public int TotalPageCount => (int)Math.Ceiling((double)TotalItemCount / PageItemSize);
+        public long TotalItemCount { get; set; }
+        public long TotalPageCount => (long)Math.Ceiling((double)TotalItemCount / PageItemSize);
 
     }
 }
